+++ routes/auth.js
@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 
 const register = require('../controllers/auth')
-const validateRegister = require('../middlewares/auth')
+const validateRegister = require('../middlewares/validations/auth')
 
 router.post('/register', validateRegister, register)
