+++ routes/categories.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const { create } = require('../controllers/categoriesController'); 
-const { validateCreate } = require('../validators/categories');
+const { validateCreate } = require('../middlewares/validations/categories');
 
 //crear usuario
 router.post('/',validateCreate, create);
