+++ routes/news.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 const { controllerAddNews } = require('../controllers/controllerNews');
-const ValidateInputNews = require('../middlewares/validations');
+const ValidateInputNews = require('../middlewares/validations/news');
 /* POST news. */
 router.post('/', ValidateInputNews,controllerAddNews);
